8/4byte integer type miss-matches
On some architectures (e.g. x86_64) ssize_t and long are 8 bytes wide
whereas int is only 4 bytes wide. This can cause at best unexpected
negative or truncated values and at worse stack corruption.

Thanks to Sergio Gelato for pointing this out.

Cc: Sergio Gelato <Sergio.Gelato@astro.su.se>

+++ perdition/io.c
@@ -655,7 +655,7 @@ err:
  *         0 otherwise (one of io_a or io_b closes gracefully)
  **********************************************************************/
 
-static int __io_pipe_read(int fd, void *buf, size_t count, void *data){
+static ssize_t __io_pipe_read(int fd, void *buf, size_t count, void *data){
   io_t *io;
   io_select_t *s;
   ssize_t bytes;
@@ -693,7 +693,9 @@ err:
 }
          
 
-static int __io_pipe_write(int fd, const void *buf, size_t count, void *data){
+static ssize_t
+__io_pipe_write(int fd, const void *buf, size_t count, void *data)
+{
   io_t *io;
   io_select_t *s;
